ReverseProxyServlet and jetty.proxy are no longer used
