+++ src/main/java/logbook/internal/proxy/PassiveModeServlet.java
@@ -5,17 +5,17 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import logbook.internal.LoggerHolder;
 import logbook.internal.ThreadManager;
-import logbook.internal.proxy.ReverseProxyServlet.RequestMetaDataWrapper;
-import logbook.internal.proxy.ReverseProxyServlet.ResponseMetaDataWrapper;
 import logbook.plugin.PluginServices;
 import logbook.proxy.ContentListenerSpi;
+import logbook.proxy.RequestMetaData;
+import logbook.proxy.ResponseMetaData;
 
 import java.io.*;
-import java.util.Base64;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Optional;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.*;
 import java.util.stream.Collectors;
+import java.util.zip.GZIPInputStream;
 
 /**
  * 外部からのHTTP POSTリクエストを受け付けてListenerを呼び出すサーブレット
@@ -150,6 +150,190 @@ public final class PassiveModeServlet extends HttpServlet {
 
         ResponseMetaDataWrapper getResponseMetaDataWrapper() {
             return this.res;
+        }
+    }
+
+    static class RequestMetaDataWrapper implements RequestMetaData, Cloneable {
+
+        private String contentType;
+
+        private String method;
+
+        private Map<String, List<String>> parameterMap;
+
+        private String queryString;
+
+        private String requestURI;
+
+        private Optional<InputStream> requestBody;
+
+        @Override
+        public String getContentType() {
+            return this.contentType;
+        }
+
+        void setContentType(String contentType) {
+            this.contentType = contentType;
+        }
+
+        @Override
+        public String getMethod() {
+            return this.method;
+        }
+
+        void setMethod(String method) {
+            this.method = method;
+        }
+
+        @Override
+        public Map<String, List<String>> getParameterMap() {
+            return this.parameterMap;
+        }
+
+        void setParameterMap(Map<String, List<String>> parameterMap) {
+            this.parameterMap = parameterMap;
+        }
+
+        @Override
+        public String getQueryString() {
+            return this.queryString;
+        }
+
+        void setQueryString(String queryString) {
+            this.queryString = queryString;
+        }
+
+        @Override
+        public String getRequestURI() {
+            return this.requestURI;
+        }
+
+        void setRequestURI(String requestURI) {
+            this.requestURI = requestURI;
+        }
+
+        @Override
+        public Optional<InputStream> getRequestBody() {
+            return this.requestBody;
+        }
+
+        void setRequestBody(Optional<InputStream> requestBody) {
+            this.requestBody = requestBody;
+        }
+
+        void set(HttpServletRequest req) {
+            this.setContentType(req.getContentType());
+            this.setMethod(req.getMethod().toString());
+            this.setQueryString(req.getQueryString());
+            this.setRequestURI(req.getRequestURI());
+        }
+
+        void set(InputStream body) {
+            String bodystr;
+            try (Reader reader = new InputStreamReader(body, StandardCharsets.UTF_8)) {
+                int len;
+                char[] cbuf = new char[128];
+                StringBuilder sb = new StringBuilder();
+                while ((len = reader.read(cbuf)) > 0) {
+                    sb.append(cbuf, 0, len);
+                }
+                bodystr = URLDecoder.decode(sb.toString(), "UTF-8");
+            } catch (IOException e) {
+                bodystr = "";
+            }
+            Map<String, List<String>> map = new LinkedHashMap<>();
+            for (String part : bodystr.split("&")) {
+                String key;
+                String value;
+                int idx = part.indexOf('=');
+                if (idx > 0) {
+                    key = part.substring(0, idx);
+                    value = part.substring(idx + 1);
+                } else {
+                    key = part;
+                    value = null;
+                }
+                map.computeIfAbsent(key, k -> new ArrayList<>())
+                        .add(value);
+            }
+            this.setParameterMap(map);
+            this.setRequestBody(Optional.of(body));
+        }
+
+        @Override
+        public RequestMetaDataWrapper clone() {
+            RequestMetaDataWrapper clone = new RequestMetaDataWrapper();
+            clone.setContentType(this.getContentType());
+            clone.setMethod(this.getMethod());
+            clone.setQueryString(this.getQueryString());
+            clone.setRequestURI(this.getRequestURI());
+            clone.setParameterMap(this.getParameterMap());
+            clone.setRequestBody(this.getRequestBody());
+            return clone;
+        }
+    }
+
+    static class ResponseMetaDataWrapper implements ResponseMetaData, Cloneable {
+
+        private int status;
+
+        private String contentType;
+
+        private Optional<InputStream> responseBody;
+
+        @Override
+        public int getStatus() {
+            return this.status;
+        }
+
+        void setStatus(int status) {
+            this.status = status;
+        }
+
+        @Override
+        public String getContentType() {
+            return this.contentType;
+        }
+
+        void setContentType(String contentType) {
+            this.contentType = contentType;
+        }
+
+        @Override
+        public Optional<InputStream> getResponseBody() {
+            return this.responseBody;
+        }
+
+        void setResponseBody(Optional<InputStream> responseBody) {
+            this.responseBody = responseBody;
+        }
+
+        void set(HttpServletResponse res) {
+            this.setStatus(res.getStatus());
+            this.setContentType(res.getContentType());
+        }
+
+        void set(InputStream body) throws IOException {
+            this.setResponseBody(Optional.of(ungzip(body)));
+        }
+
+        @Override
+        public ResponseMetaDataWrapper clone() {
+            ResponseMetaDataWrapper clone = new ResponseMetaDataWrapper();
+            clone.setStatus(this.getStatus());
+            clone.setContentType(this.getContentType());
+            clone.setResponseBody(this.getResponseBody());
+            return clone;
+        }
+
+        private static InputStream ungzip(InputStream body) throws IOException {
+            body.mark(Short.BYTES);
+            int magicbyte = body.read() << 8 ^ body.read();
+            body.reset();
+            if (magicbyte == 0x1f8b) {
+                return new GZIPInputStream(body);
+            }
+            return body;
         }
     }
 }
